move config package to internal
