+++ internal/restapi/v0/scan/post.go
@@ -9,8 +9,8 @@ import (
 	_ "strings"
 	"time"
 
+	"github.com/ashwiniag/goKakashi/internal/config/v0"
 	"github.com/ashwiniag/goKakashi/internal/imgscan"
-	"github.com/ashwiniag/goKakashi/pkg/config/v0"
 	_ "github.com/ashwiniag/goKakashi/pkg/utils"
 	"github.com/scriptnull/jsonseal"
 	"github.com/swaggest/usecase/status"
