+++ pkg/registry/dockerhub.go
@@ -6,7 +6,7 @@ import (
 	"os/exec"
 	"strings"
 
-	"github.com/ashwiniag/goKakashi/pkg/config/v0"
+	"github.com/ashwiniag/goKakashi/internal/config/v0"
 )
 
 type DockerHub struct{}
